+++ db/migrate/20180711132009_add_rating_to_service.rb
@@ -1,0 +1,5 @@
+class AddRatingToService < ActiveRecord::Migration[5.2]
+  def change
+    add_column :services, :rating, :decimal,  default: 0.0, null: false, precision: 2, scale: 1
+  end
+end