+++ db/schema.rb
@@ -10,8 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-
-ActiveRecord::Schema.define(version: 2018_07_06_101250) do
+ActiveRecord::Schema.define(version: 2018_07_11_132009) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -85,6 +84,7 @@ ActiveRecord::Schema.define(version: 2018_07_06_101250) do
     t.text "terms_of_use"
     t.text "tagline", null: false
     t.bigint "owner_id"
+    t.decimal "rating", precision: 2, scale: 1, default: "0.0", null: false
     t.index ["description"], name: "index_services_on_description"
     t.index ["owner_id"], name: "index_services_on_owner_id"
     t.index ["title"], name: "index_services_on_title"