+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Service and categories database seed (@bwilk)
 - Conditional software version placeholder in frontend (@bwilk)
 - Service owner can manage owned services (@mkasztelnik)
+- Add rating for Service (@kmarszalek)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)